+++ crates/sncast/src/response/utils/mod.rs
@@ -1,2 +1,3 @@
 pub mod class_hash;
+pub mod contract_address;
 pub mod serialize;