+++ crates/sncast/src/starknet_commands/utils/mod.rs
@@ -15,11 +15,14 @@ use crate::{
     process_command_result,
     starknet_commands::{
         self,
-        utils::{class_hash::ClassHash, serialize::Serialize},
+        utils::{
+            class_hash::ClassHash, contract_address::ContractAddress, serialize::Serialize,
+        },
     },
 };
 
 pub mod class_hash;
+pub mod contract_address;
 pub mod felt_or_id;
 pub mod serialize;
 
@@ -36,6 +39,9 @@ pub enum Commands {
 
     /// Get contract class hash
     ClassHash(ClassHash),
+
+    /// Calculate the address of a not yet deployed contract
+    ContractAddress(ContractAddress),
 }
 
 pub async fn utils(
@@ -75,6 +81,39 @@ pub async fn utils(
                 .map_err(handle_starknet_command_error)?;
 
             process_command_result("class-hash", Ok(result), ui, None);
+        }
+
+        Commands::ContractAddress(contract_address) => {
+            let artifacts =
+                if contract_address.common.contract_identifier.contract_name.is_some() {
+                    let manifest_path = assert_manifest_path_exists()?;
+                    let package_metadata =
+                        get_package_metadata(&manifest_path, &contract_address.common.package)?;
+
+                    Some(
+                        build_and_load_artifacts(
+                            &package_metadata,
+                            &BuildConfig {
+                                scarb_toml_path: manifest_path,
+                                json,
+                                profile,
+                            },
+                            false,
+                            // TODO(#3959) Remove `base_ui`
+                            ui.base_ui(),
+                        )
+                        .expect("Failed to build contract"),
+                    )
+                } else {
+                    None
+                };
+
+            let result =
+                contract_address::get_contract_address(contract_address, artifacts, config, ui)
+                    .await
+                    .map_err(handle_starknet_command_error)?;
+
+            process_command_result("contract-address", Ok(result), ui, None);
         }
     }